+++ kikuchipy/indexing/tests/test_merge_crystal_maps.py
@@ -17,6 +17,8 @@
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
+from orix.crystal_map import CrystalMap, Phase, PhaseList
+from orix.quaternion import Rotation
 import pytest
 
 from kikuchipy.indexing._merge_crystal_maps import merge_crystal_maps
@@ -253,7 +255,9 @@ class TestMergeCrystalMaps:
         desired_phase_id[3] = 1
 
         merged_xmap = merge_crystal_maps(
-            crystal_maps=[xmap1, xmap2], metric=metric,
+            crystal_maps=[xmap1, xmap2],
+            metric=metric,
+            simulation_indices_prop=sim_idx_prop,
         )
 
         assert np.allclose(merged_xmap.phase_id, desired_phase_id)
@@ -356,10 +360,13 @@ class TestMergeCrystalMaps:
         """Ensure that the mergesorted scores and simulation index
         properties in the merged map has the correct values and shape.
         """
+        prop_names = ["scores", "simulation_indices"]
         n_phases = np.shape(desired_merged_scores)[-1] // rot_per_point
         xmaps = []
         for i in range(n_phases):
-            xmap = get_single_phase_xmap(nav_shape, rot_per_point, name=str(i))
+            xmap = get_single_phase_xmap(
+                nav_shape, rot_per_point, name=str(i), prop_names=prop_names
+            )
             xmap[i].scores += i
             xmaps.append(xmap)
 
@@ -368,7 +375,9 @@ class TestMergeCrystalMaps:
         assert np.sum(np.diff(all_sim_idx)) == 0
 
         merged_xmap = merge_crystal_maps(
-            crystal_maps=xmaps, mean_n_best=mean_n_best,
+            crystal_maps=xmaps,
+            mean_n_best=mean_n_best,
+            simulation_indices_prop=prop_names[1],
         )
 
         assert merged_xmap.phases.size == n_phases
@@ -415,3 +424,51 @@ class TestMergeCrystalMaps:
         crystal_maps = [xmap1, xmap2]
         with pytest.raises(ValueError, match="All crystal maps must have the"):
             _ = merge_crystal_maps(crystal_maps)
+
+    def test_merging_refined_maps(self):
+        ny, nx = (3, 3)
+        nav_size = ny * nx
+        r = Rotation.from_euler(np.ones((nav_size, 3)))
+        x = np.tile(np.arange(ny), nx)
+        y = np.repeat(np.arange(nx), ny)
+
+        # Simulation indices
+        n_sim_indices = 10
+        sim_indices1 = np.random.randint(
+            low=0, high=1000, size=n_sim_indices * nav_size
+        ).reshape((nav_size, n_sim_indices))
+        sim_indices2 = np.random.randint(
+            low=0, high=1000, size=n_sim_indices * nav_size
+        ).reshape((nav_size, n_sim_indices))
+
+        # Scores
+        scores1 = np.ones(nav_size)
+        scores1[0] = 3
+        scores2 = 2 * np.ones(nav_size)
+
+        xmap1 = CrystalMap(
+            rotations=r,
+            phase_id=np.ones(nav_size) * 0,
+            phase_list=PhaseList(Phase(name="a")),
+            x=x,
+            y=y,
+            prop={"simulation_indices": sim_indices1, "scores": scores1},
+        )
+        xmap2 = CrystalMap(
+            rotations=r,
+            phase_id=np.ones(nav_size),
+            phase_list=PhaseList(Phase(name="b")),
+            x=x,
+            y=y,
+            prop={"simulation_indices": sim_indices2, "scores": scores2},
+        )
+        xmap_merged = merge_crystal_maps(crystal_maps=[xmap1, xmap2])
+
+        assert "simulation_indices" not in xmap_merged.prop.keys()
+        assert "merged_simulation_indices" not in xmap_merged.prop.keys()
+
+        with pytest.raises(ValueError, match="Cannot merge maps with more"):
+            _ = merge_crystal_maps(
+                crystal_maps=[xmap1, xmap2],
+                simulation_indices_prop="simulation_indices",
+            )