+++ doc/changelog.rst
@@ -33,7 +33,7 @@ Added
 Fixed
 -----
 - Merging of (typically refined) crystal maps, where either a simulation indices
-  is not present, or it contains more indices per point than scores.
+  array is not present or the array contains more indices per point than scores.
   (`#335 <https://github.com/pyxem/kikuchipy/pull/335>`_)
 - Bugs in getting plot markers from geometrical EBSD simulation.
   (`#334 <https://github.com/pyxem/kikuchipy/pull/334>`_)