+++ doc/changelog.rst
@@ -24,7 +24,7 @@ Fixed
 -----
 - Merging of (typically refined) crystal maps, where either a simulation indices
   is not present, or it contains more indices per point than scores.
-  (`#? <https://github.com/pyxem/kikuchipy/pull/?>`_)
+  (`#335 <https://github.com/pyxem/kikuchipy/pull/335>`_)
 - Bugs in getting plot markers from geometrical EBSD simulation.
   (`#334 <https://github.com/pyxem/kikuchipy/pull/334>`_)
 - Passing a static background pattern to EBSD.remove_static_background() for a